+++ include/core/traceable_item.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "contracts/traceable_item.h"